feat: implement data validation and deduplication.
Data verification scenario: API service when putting objects and data service when getting objects.

Deduplication service refers to the singleton check during the upload process of objects, so that the contents of objects that already exist on disk will not be uploaded repeatedly, but new version information will be recorded.