feat: implement data redundancy and instant repair.
Data Redundancy: Based on Reed-Solomon error correction coding technology, the data is divided into N data shards of equal size and M error correction shards, each of which has the same size and is one N of the total size. Any N shards can restore the original data.

Instant Data Repair:When the user tries to get data, RSGetStream will check the sharding situation in the getData() function, and if there is no exception, the data can be returned normally. If a shard is lost, the content of the shard is wrong due to damage or tampering (the abnormal shard will be cleared), these conditions cannot be read normally, then d.enc.Reconstruct(shards) will be called to try to repair the abnormal shard, and the data can be returned after all the shards are restored to normal. This process requires at least N healthy shards.