feat: implement data compression.
In the PUT process: when the temporary data shards are normalized, they were originally directly renamed, but now they are first gzip compressed and then renamed to the official shard file, so the shards saved on the data service nodes are compressed.

In the GET process: the data service decompresses the gzip compressed shards and then passes them to the API service, which returns the gzip file or the original file according to the customer request header.