feat: implement data maintenance.
deleteOldMetadata: Set the metadata version retention policy, here we limit each object to keep only the latest 5 versions, if the limit is exceeded, the oldest metadata will be deleted.

deleteOrphanObject: When outdated metadata versions are purged or due to some other anomaly, unreferenced objects will appear, which will be periodically purged by deleteOrphanObject.go (these objects will be moved to the garbage directory for further processing at the discretion of the administrator, and are usually deleted completely periodically).

objectScanner: Periodically, the tool checks and repairs all objects in the data node, and if there are too few normal shards, it cannot be recovered.